+++ app/helpers/hedgedocs_helper.rb
@@ -1,19 +1,19 @@
-module CodimdsHelper
-  def new_codimd_pad_path
-    "#{RedmineCodimd.setting :codimd_url}/new"
+module HedgedocsHelper
+  def new_hedgedoc_pad_path
+    "#{RedmineHedgedoc.setting :hedgedoc_url}/new"
   end
 
-  def codimd_pad_url(note)
-    "#{RedmineCodimd.setting :codimd_url}/#{note.shortid}"
+  def hedgedoc_pad_url(note)
+    "#{RedmineHedgedoc.setting :hedgedoc_url}/#{note.shortid}"
   end
 
-  def codimd_pad_owner(note)
+  def hedgedoc_pad_owner(note)
     return if note.User.blank? || note.User.email.blank?
 
     User.having_mail(note.User.email).first
   end
 
-  def codimd_pad_link(note, project = nil)
+  def hedgedoc_pad_link(note, project = nil)
     title = if project.present?
               len = project.identifier.length + 1
               max_len = note.title.length
@@ -21,6 +21,6 @@ module CodimdsHelper
             else
               note.title
             end
-    link_to h(title), codimd_pad_url(note), class: 'external'
+    link_to h(title), hedgedoc_pad_url(note), class: 'external'
   end
 end