+++ app/models/hedgedoc_user.rb
@@ -1,5 +1,5 @@
-class CodimdUser < ActiveRecord::Base
-  include CodimdDatabase
+class HedgedocUser < ActiveRecord::Base
+  include HedgedocDatabase
   self.table_name = 'Users'
 
   has_many :Notes, dependent: :destroy, foreign_key: 'id', inverse_of: :ownerId