+++ test/integration/routing_test.rb
@@ -1,8 +1,8 @@
 require File.expand_path '../../test_helper', __FILE__
 
 class RoutingTest < Redmine::RoutingTest
-  test 'codimd' do
-    should_route 'GET /codimd' => 'codimds#show'
-    should_route 'GET /projects/foo/codimd' => 'codimds#show', project_id: 'foo'
+  test 'hedgedoc' do
+    should_route 'GET /hedgedoc' => 'hedgedocs#show'
+    should_route 'GET /projects/foo/hedgedoc' => 'hedgedocs#show', project_id: 'foo'
   end
 end