+++ test/test_helper.rb
@@ -15,25 +15,25 @@ end
 
 require File.expand_path "#{File.dirname __FILE__}/../../../test/test_helper"
 
-module RedmineCodimd
+module RedmineHedgedoc
   module TestHelper
     def prepare_tests
       Role.where(id: [1, 2]).each do |r|
-        r.permissions << :show_codimd_pads
+        r.permissions << :show_hedgedoc_pads
         r.save
       end
 
       Project.where(id: [1, 2]).each do |project|
-        EnabledModule.create(project: project, name: 'codimd')
+        EnabledModule.create project: project, name: 'hedgedoc'
       end
     end
   end
 
   class ControllerTest < Redmine::ControllerTest
-    include RedmineCodimd::TestHelper
+    include RedmineHedgedoc::TestHelper
   end
 
   class TestCase < ActiveSupport::TestCase
-    include RedmineCodimd::TestHelper
+    include RedmineHedgedoc::TestHelper
   end
 end