+++ test/unit/i18n_test.rb
@@ -1,6 +1,6 @@
 require File.expand_path '../../test_helper', __FILE__
 
-class I18nTest < RedmineCodimd::TestCase
+class I18nTest < RedmineHedgedoc::TestCase
   include Redmine::I18n
 
   def setup
@@ -17,16 +17,16 @@ class I18nTest < RedmineCodimd::TestCase
   end
 
   def test_locales_validness
-    lang_files_count = Dir[Rails.root.join('plugins/redmine_codimd/config/locales/*.yml')].size
+    lang_files_count = Dir[Rails.root.join('plugins/redmine_hedgedoc/config/locales/*.yml')].size
     assert_equal lang_files_count, 2
     valid_languages.each do |lang|
       assert set_language_if_valid(lang)
     end
     # check if parse error exists
     ::I18n.locale = 'de'
-    assert_equal 'Neues CodiMD Pad', l(:label_new_codimd_pad)
+    assert_equal 'Neues HedgeDoc Pad', l(:label_new_hedgedoc_pad)
     ::I18n.locale = 'en'
-    assert_equal 'New CodiMD pad', l(:label_new_codimd_pad)
+    assert_equal 'New HedgeDoc pad', l(:label_new_hedgedoc_pad)
 
     set_language_if_valid('en')
   end